izmjene

+++ watchlist-service/watchlist-service/src/main/java/com/moviehub/watchlistservice/POJO/Actor/AddMovieForActorRequest.java
@@ -4,5 +4,5 @@ public record AddMovieForActorRequest(
         Long movieId,
         String name,
         String textDescription,
-        Double grade
+        Float grade
 ){}
