+++ watchlist-service/watchlist-service/src/main/java/com/moviehub/watchlistservice/POJO/Genre/AddMovieForGenreRequest.java
@@ -4,6 +4,6 @@ public record AddMovieForGenreRequest(
         Long movieId,
         String name,
         String textDescription,
-        Double grade
+        Float grade
 ) {
 }
