+++ watchlist-service/watchlist-service/src/main/java/com/moviehub/watchlistservice/POJO/Watchlist/AddMovieToWatchlistRequest.java
@@ -4,6 +4,6 @@ public record AddMovieToWatchlistRequest(
         Long movieId,
         String name,
         String textDescription,
-        Double grade
+        Float grade
 ) {
 }
